+++ helpers/ingredients.js
@@ -9,8 +9,8 @@ var UserHelper = require('./users');
 var Spending = require('../models/spending');
 var Production = require('../models/production');
 var Freshness = require('../models/freshness');
+var Completed = require('../models/completed_production');
 var Recall = require('../models/recall');
-var RealRecall = require('../models/real_recall');
 var mongoose = require('mongoose');
 mongoose.Promise = global.Promise;
 
@@ -133,13 +133,13 @@ module.exports.decrementAmountForProduction = function(id, amount, formula, form
     }).then(async function(results) {
       let lotsConsumed = results[1];
       for (let lot of lotsConsumed) {
-        await RealRecall.createLotEntry(ing._id, ing.name, lot.lotNumber, lot.vendorID);
+        await Recall.createLotEntry(ing._id, ing.name, lot.lotNumber, lot.vendorID);
       }
       for (let lot of lotsConsumed) {
         var time = Date.now() - lot.timestamp;
         await Freshness.updateReport(lot.ingID, lot.name, lot.amount, time);
-        await Recall.updateReport(formula.name, formulaLot, lot.ingID, lot.lotNumber, lot.vendorID);
-        await RealRecall.updateReport(formula.id, formulaLot, formula.intermediate, lot.ingID, lot.lotNumber, lot.vendorID);
+        await Completed.updateReport(formula.name, formulaLot, lot.ingID, lot.lotNumber, lot.vendorID);
+        await Recall.updateReport(formula.id, formulaLot, formula.intermediate, lot.ingID, lot.lotNumber, lot.vendorID);
       }
       return "done";
     }).then(function(result) {
