+++ models/recall.js
@@ -1,6 +1,7 @@
 var mongoose = require('mongoose');
 var Ingredient = require('./ingredient');
 var Vendor = require('./vendor');
+var Formula = require('./formula');
 mongoose.Promise = global.Promise;
 
 var RecallSchema = new mongoose.Schema({
@@ -9,18 +10,19 @@ var RecallSchema = new mongoose.Schema({
     required: true,
     unique: true
   },
-  products: [
+  vendorLots: [
     {
-      name: String,
+      vendorID: String,
+      vendorName: String,
+      vendorCode: String,
       lotNumber: String,
-      intermediate: Boolean,
-      constituents: [
+      ingredientName: String,
+      products: [
         {
-          ingredientID: String,
-          ingredientName: String,
+          intermediate: Boolean,
+          productName: String,
           lotNumber: String,
-          vendorID: String,
-          vendorName: String
+          formulaID: String
         }
       ]
     }
@@ -31,14 +33,14 @@ var Recall = mongoose.model('Recall', RecallSchema);
 
 module.exports.model = Recall;
 
-// checkNewLot = function(formulaName, formulaLot, report) {
-//   for (let record of report['products']) {
-//     if (record['name'] == formulaName && record['lotNumber'] == formulaLot) {
-//       return false;
-//     }
-//   }
-//   return true;
-// }
+module.exports.checkNewLot = function(vendorID, lotNumber, report) {
+  for (let record of report['vendorLots']) {
+    if (record['vendorID'] == vendorID && record['lotNumber'] == lotNumber) {
+      return false;
+    }
+  }
+  return true;
+}
 
 createReport = function() {
   return new Promise(function(resolve, reject) {
@@ -46,7 +48,7 @@ createReport = function() {
       if (report == null) {
         resolve(Recall.create({
           'name': 'recall',
-          'products': []
+          'vendorLots': []
         }));
       } else {
         resolve(report);
@@ -57,33 +59,71 @@ createReport = function() {
   });
 }
 
-module.exports.createLotEntry = function(formulaName, formulaLot, intermediate) {
-  let newEntry = {
-    'name': formulaName,
-    'lotNumber': formulaLot,
-    'intermediate': intermediate,
-    'constituents': []
-  };
-  return Recall.findOneAndUpdate({'name': 'recall'}, {'$push': {'products': newEntry}}).exec();
+module.exports.createLotEntry = function(ingID, ingName, ingLot, vendorID) {
+  return new Promise(function(resolve, reject) {
+    var newEntry;
+    var promise;
+    if (vendorID == 'admin') {
+      promise = new Promise(function(resolve, reject) {
+        newEntry = {
+          'vendorID': vendorID,
+          'vendorName': vendorID,
+          'vendorCode': vendorID,
+          'lotNumber': ingLot,
+          'ingredientName': ingName,
+          'products': []
+        };
+        resolve(newEntry);
+      });
+    } else {
+      promise = new Promise(function(resolve, reject) {
+        Vendor.findVendorById(vendorID).then(function(vendor) {
+          newEntry = {
+            'vendorID': vendorID,
+            'vendorName': (vendor == null) ? vendorID : vendor.name,
+            'vendorCode': (vendor == null) ? vendorID : vendor.code,
+            'lotNumber': ingLot,
+            'ingredientName': ingName,
+            'products': []
+          };
+          resolve(newEntry);
+        }).catch(function(error) {
+          reject(error);
+        });
+      });
+    }
+    promise.then(function(entry) {
+      return Recall.findOne({'$and': [{'name': 'recall'}, {'vendorLots': {'$elemMatch': {'vendorID': vendorID, 'lotNumber': ingLot, 'ingredientName': ingName}}}]});
+    }).then(function(recall) {
+      if (recall == null) {
+        resolve(Recall.findOneAndUpdate({'name': 'recall'}, {'$push': {'vendorLots': newEntry}}).exec());
+      } else {
+        resolve(recall);
+      }
+    }).catch(function(error) {
+      reject(error);
+    })
+  })
 }
 
-module.exports.updateReport = function(formulaName, formulaLot, ingID, ingLot, vendorID) {
+module.exports.updateReport = function(formulaID, formulaLot, intermediate, ingID, ingLot, vendorID) {
   return new Promise(function(resolve, reject) {
+    var globalReport;
     createReport().then(function(result) {
       return Recall.findOne( {'name': 'recall'} ).exec();
     }).then(function(report) {
+      globalReport = report;
       if (vendorID == 'admin') {
-        return Promise.all([Ingredient.getIngredientById(ingID), {'name': 'admin'}]);
+        return Promise.all([Ingredient.getIngredientById(ingID), {'name': 'admin'}, Formula.model.findById(formulaID).exec()]);
       } else {
-        return Promise.all([Ingredient.getIngredientById(ingID), Vendor.findVendorById(vendorID)]);
+        return Promise.all([Ingredient.getIngredientById(ingID), Vendor.findVendorById(vendorID), Formula.model.findById(formulaID).exec()]);
       }
     }).then(function(result) {
       var ing = result[0];
       var vendor = result[1];
-      let newConstituent = {'ingredientID': ingID, 'ingredientName': ing.name, 'lotNumber': ingLot, 'vendorID': vendorID, 
-      'vendorName': vendor.name};
-      return Recall.update({'$and': [{'name': 'recall'}, {'products': {'$elemMatch': {'name': formulaName, 'lotNumber': formulaLot}}}]},
-        {'$push': {'products.$.constituents': newConstituent}}).exec();
+      var formula = result[2];
+      let newProduct = {'intermediate': intermediate, 'productName': formula.name, 'lotNumber': formulaLot, 'formulaID': formula._id};
+      return Recall.update({'$and': [{'name': 'recall'}, {'vendorLots': {'$elemMatch': {'vendorID': vendorID, 'lotNumber': ingLot, 'ingredientName': ing.name}}}]}, {'$push': {'vendorLots.$.products': newProduct}}).exec();
     }).then(function(report) {
       resolve(report);
     }).catch(function(error) {
@@ -92,15 +132,55 @@ module.exports.updateReport = function(formulaName, formulaLot, ingID, ingLot, v
   })
 }
 
-module.exports.getProducts = function() {
+checkIfProductInArray = function(finalProducts, product) {
+  for (let record in finalProducts) {
+    if (record['productName'] == product['name'] && record['lotNumber'] == product['lotNumber']) {
+      return true;
+    }
+  }
+  return false;
+}
+
+module.exports.getRecallProducts = function(ingName, vendorCode, lotNumber) {
   return new Promise(function(resolve, reject) {
-    Recall.findOne({'name': 'recall'}).then(function(report) {
-      let returnProducts = [];
-      for (let record of report.products) {
-        record['timestamp'] = mongoose.Types.ObjectId(record['_id']).getTimestamp().toString();
-        returnProducts.push(record);
+    let finalProducts = [];
+    exports.getRecall().then(function(report) {
+      let products = [];
+      for (let record of report['vendorLots']) {
+        if (record['vendorCode'] == vendorCode && record['lotNumber'] == lotNumber && record['ingredientName'] == ingName) {
+          products = record.products;
+          break;
+        }
       }
-      resolve(returnProducts);
+      let intermediates = [];
+      for (let product of products) {
+        if (product.intermediate) {
+          intermediates.push({'name': product.productName, 'vendorCode': 'admin', 'lotNumber': product.lotNumber});
+        } else {
+          // if (!checkIfProductInArray(finalProducts, product)) {
+          finalProducts.push(product);
+          // }
+        }
+      }
+      if (intermediates.length == 0) {
+        return [];
+      } else {
+        console.log(intermediates);
+        return Promise.all(intermediates.map(function(tuple) {
+          return exports.getRecallProducts(tuple.name, tuple.vendorCode, tuple.lotNumber);
+        }));
+      }
+    }).then(function(results) {
+      console.log('reeeeeeeeee');
+      console.log(finalProducts);
+      results.forEach(function(arr) {
+        for (let product of arr) {
+          if (!checkIfProductInArray(finalProducts, product)) {
+            finalProducts.push(product);
+          }
+        }
+      });
+      resolve(finalProducts);
     }).catch(function(error) {
       reject(error);
     })
