+++ package.json
@@ -38,6 +38,7 @@
     "morgan": "~1.9.0",
     "node-schedule": "^1.3.0",
     "nodemailer": "^4.4.2",
+    "npm": "^5.8.0",
     "oauth": "^0.9.15",
     "papaparse": "^4.3.7",
     "popups": "^1.1.3",
